+++ bootstrap/development.js
@@ -1,6 +1,5 @@
 const { join } = require('path');
-import packageJson from '../package';
 require('dotenv').config();
 
 // Started from /src directory
-require(join(process.cwd(), 'src', packageJson.config["app-name"]));
+require(join(process.cwd(), 'src'));
