+++ bootstrap/production.js
@@ -1,5 +1,4 @@
 const { join } = require('path');
-import packageJson from '../package';
 
 // From the built source code
-require(join(process.cwd(), 'dist', packageJson.config["app-name"]));
+require(join(process.cwd(), 'dist'));
