+++ package.json
@@ -5,9 +5,6 @@
   "main": "./src/sample-service/index.js",
   "author": "Theekshana Wijesinghe",
   "license": "MIT",
-  "config": {
-    "app-name": "sample-service"
-  },
   "scripts": {
     "build": "NODE_ENV=production yarn clean && babel src -d dist",
     "start": " NODE_ENV=development nodemon --exec babel-node ./bootstrap",
